contour: fix the Windows and macOS builds
Two build breaks, both found by CI reaching platforms this branch could not be
compiled on locally.

macOS: CocoaWindowFrame.mm still named TitleBarDecoration, which was renamed to
WindowDecoration during the simplify pass -- the rename swept .cpp and .hpp and
never touched the tree's only .mm. The stale parameter type left the class
abstract, so the error surfaced twice over.

Windows: Dwmapi.hpp used MARGINS with only <Windows.h> included. It is declared
in <uxtheme.h>, which <dwmapi.h> pulls in. Including them declares the DWM entry
points as well, which is harmless -- they are still resolved at runtime, so the
declarations serve only as types and dwmapi.lib stays out of the link line.
Also spells out the DWORD conversion on two sizeof arguments, which -Werror
rejects on the Windows toolchains.

This commit also pinned the Linux Qt install to 6.11.1, around 6.11.2's broken
Linux packages. That hunk is dropped here: master now pins every platform to
6.11.1 and retries the install past a bad mirror, so the workflow already
carries the fix and this branch no longer needs its own.

+++ src/contour/platform/Win32WindowFrame.cpp
@@ -213,11 +213,14 @@ namespace
             // The border DWM draws around the rounded corners is tinted by this. Driven from the
             // CONFIGURED theme: it used to be forced on in BlurBehind.cpp regardless, which put a
             // dark border on a light-themed window.
+            // sizeof yields size_t and the attribute size is a DWORD; spelled out rather than left
+            // to an implicit narrowing, which -Werror rejects on the Windows toolchains.
             auto dark = static_cast<BOOL>(_colorScheme() == Qt::ColorScheme::Dark);
-            api.setWindowAttribute(hwnd, DwmwaUseImmersiveDarkMode, &dark, sizeof(dark));
+            api.setWindowAttribute(hwnd, DwmwaUseImmersiveDarkMode, &dark, static_cast<DWORD>(sizeof(dark)));
 
             auto corner = DwmwcpRound;
-            api.setWindowAttribute(hwnd, DwmwaWindowCornerPreference, &corner, sizeof(corner));
+            api.setWindowAttribute(
+                hwnd, DwmwaWindowCornerPreference, &corner, static_cast<DWORD>(sizeof(corner)));
         }
 
         void invalidateShadow(QWindow& /*window*/) override
